was not careful

+++ openbb_platform/openbb/package/equity_ownership.py
@@ -116,7 +116,7 @@ class ROUTER_equity_ownership(Container):
             Expiration date of the derivative. (provider: intrinio)
         underlying_security_title : Optional[str]
             Name of the underlying non-derivative security related to this derivative transaction. (provider: intrinio)
-        underlying_shares : Optional[Union[int, float]]
+        underlying_shares : Optional[Union[float, int]]
             Number of underlying shares related to this derivative transaction. (provider: intrinio)
         nature_of_ownership : Optional[str]
             Nature of ownership of the insider trading. (provider: intrinio)
